src/requirements.txt
absl-py==0.4.0
astor==0.6.2
bleach==1.5.0
certifi==2018.8.13
chardet==3.0.4
click==6.7
colorama==0.3.9
cycler==0.10.0
decorator==4.1.2
dev==0.4.0
enum34==1.1.6
Flask==1.0.2
ftfy==5.4.1
gast==0.2.0
grpcio==1.11.0
gunicorn==19.9.0
h5py==2.7.1
html5lib==0.9999999
idna==2.7
ipython==6.2.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.11.0
Jinja2==2.10
Keras==2.1.6
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.0
mlxtend==0.9.1
nltk==3.3
nodejs==0.1.1
npm==0.1.1
numpy==1.15.0
opencv-python==3.4.2.17
optional-django==0.1.0
pandas==0.21.0
parso==0.1.0
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
protobuf==3.6.1
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-resize-image==1.1.11
pytz==2017.3
PyYAML==3.12
requests==2.19.1
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
tensorboard==1.10.0
tensorflow==1.10.0
tensorflow-gpu==1.9.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
traitlets==4.3.2
urllib3==1.23
virtualenv==16.0.0
waitress==1.1.0
wcwidth==0.1.7
Werkzeug==0.14.1
win-unicode-console==0.5
